Added "addNewLesson" function to a new file
Also updated the "mapTimeToIndex" Function
