+++ src/views/CoursesPanel/CoursesPanel_Functions.jsx
@@ -1,79 +1,23 @@
-import Course from './CoursesPanel';
-import { apis } from '../../assets/apis';
-// Token
-// const tokenJson = localStorage.getItem("authTokens");
-// const tokenClass = JSON.parse(tokenJson);
-// const token = tokenClass.token.access;
+export function mapTimeToIndex(startTime) {
+  const timeRanges = [
+    ["07:30:00", "09:00:00"],
+    ["09:00:01", "10:30:00"],
+    ["10:30:01", "12:00:00"],
+    ["13:00:00", "14:30:00"],
+    ["14:30:01", "16:00:00"],
+    ["16:00:01", "17:30:00"],
+    ["17:30:01", "19:00:00"],
+    ["19:00:01", "20:30:00"]
+  ];
 
-export function addNewLesson(num) {  // Add a lesson
-  const tokenJson = localStorage.getItem("authTokens");
-  const tokenClass = JSON.parse(tokenJson);
-  const token = tokenClass.token.access;
-                                     
-    fetch(apis["courses"]["my_courses"], {
-      method: "PUT",
-      headers: {
-        Authorization: `Bearer ${token}`,
-        Accept: "application/json",
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        complete_course_number: num,
-      }),
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        const courses = data.map(course => new Course(course, true));
-        // setChosenCourses(courses);
-        return courses;
-        // setChosenCoursesChanged(prev => !prev);
-      })
-      .catch((error) => console.error(error));
-    const activeRoute = (routeName) => {
-      return location.pathname === routeName ? "active" : "";
-    };
-}
-
-export function mapTimeToIndex(start_time) {                               // Time Index finder
-    const times = [
-      "07:30:00",
-      "09:00:00",
-      "10:30:00",
-      "12:00:00",
-      "13:30:00",
-      "15:00:00",
-      "16:30:00",
-      "18:00:00",
-    ];
-    let index = times.findIndex(time => time === start_time);
-    if (index === -1) {
-      const times2 = [
-        "08:00:00",
-        "09:30:00",
-        "11:00:00",
-        "12:30:00",
-        "14:00:00",
-        "15:30:00",
-        "17:00:00",
-        "18:30:00",
-      ];
-      let index2 = times2.findIndex(time => time === start_time);
-      if (index2 === -1) {
-        const times3 = [
-          "07:00:00",
-          "08:30:00",
-          "10:00:00",
-          "11:30:00",
-          "13:00:00",
-          "14:30:00",
-          "16:00:00",
-          "17:30:00",
-        ];
-        return times3.findIndex(time => time === start_time);
-      }
-      return index2;
+  for (let i = 0; i < timeRanges.length; i++) {
+    const [start, end] = timeRanges[i];
+    if (startTime >= start && startTime <= end) {
+      return i;
     }
-    return index;
+  }
+
+  return -1;
 }
 
 export function createCourseGroupsArray(courses) {                         // Creates an array to store number of courses in each time slot
